is_tuple<T> and tuple_head<T> now work not only with tuple<T...> itself but also with objects derived from tuple<T...>.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2570 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ TUTools++/Testprog/simd/transform/main.cc
@@ -71,6 +71,6 @@ doJob()
 int
 main()
 {
-    TU::simd::doJob<float, int8_t, int8_t, int16_t, int32_t>();
+    TU::simd::doJob<float, int16_t, int8_t, int16_t, int32_t>();
     return 0;
 }
